% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHPWeight.R
\name{AHPWeight}
\alias{AHPWeight}
\title{AHPWeight
Calculate weights on standardized data us the entropy method.}
\usage{
AHPWeight(matrixdata)
}
\arguments{
\item{matrixdata}{Judgment matrix data.}
}
\value{
weightResult is weight value of Index
}
\description{
AHPWeight
Calculate weights on standardized data us the entropy method.
}
\examples{
\donttest{
matrixdata<-MatixData
AHPWeight(matrixdata)
}
}
