% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatixData.R
\docType{data}
\name{MatixData}
\alias{MatixData}
\title{Judgment matrix of sampleData.}
\format{
A data fram with 11 rows and 12 variables.
}
\source{
Subjective judgment.
}
\usage{
MatixData
}
\description{
Judgment matrix of sampleData.
}
\keyword{datasets}
